Move all string-related logic into a single file.
