+++ packages/math-input/src/components/input/__tests__/context-tracking.test.ts
@@ -1,4 +1,4 @@
-import {strings} from "../../../mock-strings";
+import {mockStrings} from "../../../strings";
 import {CursorContext} from "../cursor-contexts";
 
 import TestMathWrapper from "./test-math-wrapper";
@@ -11,7 +11,7 @@ describe("Cursor context", () => {
         span = document.createElement("span");
         document.body.appendChild(span);
 
-        mathField = new TestMathWrapper(span, strings, "en");
+        mathField = new TestMathWrapper(span, mockStrings, "en");
     });
 
     afterEach(() => {
