+++ packages/math-input/src/components/input/__tests__/mathquill-helpers.test.ts
@@ -1,4 +1,4 @@
-import {strings} from "../../../mock-strings";
+import {mockStrings} from "../../../strings";
 import {CursorContext} from "../cursor-contexts";
 import {getCursorContext} from "../mathquill-helpers";
 import {createMathField} from "../mathquill-instance";
@@ -8,7 +8,7 @@ describe("MathQuill Helpers", () => {
         it("returns NONE for empty MathField", () => {
             // Arrange
             const mount = document.createElement("div");
-            const mathField = createMathField(mount, strings);
+            const mathField = createMathField(mount, mockStrings);
 
             // Act
             const context = getCursorContext(mathField);
@@ -20,7 +20,7 @@ describe("MathQuill Helpers", () => {
         it("returns BEFORE_FRACTION when before a fraction", () => {
             // Arrange
             const mount = document.createElement("div");
-            const mathField = createMathField(mount, strings);
+            const mathField = createMathField(mount, mockStrings);
             // create fraction and move cursor before it
             mathField.cmd("\\frac");
             mathField.keystroke("Left");
@@ -35,7 +35,7 @@ describe("MathQuill Helpers", () => {
         it("returns IN_PARENS when in parenthesis", () => {
             // Arrange
             const mount = document.createElement("div");
-            const mathField = createMathField(mount, strings);
+            const mathField = createMathField(mount, mockStrings);
             // MQ puts you inside parens when you start a paren
             mathField.typedText("(");
 
@@ -49,7 +49,7 @@ describe("MathQuill Helpers", () => {
         it("returns IN_NUMERATOR when in numerator", () => {
             // Arrange
             const mount = document.createElement("div");
-            const mathField = createMathField(mount, strings);
+            const mathField = createMathField(mount, mockStrings);
             // MQ puts you in the numerator when making a fraction
             mathField.cmd("\\frac");
 
@@ -63,7 +63,7 @@ describe("MathQuill Helpers", () => {
         it("returns IN_DENOMINATOR when in denominator", () => {
             // Arrange
             const mount = document.createElement("div");
-            const mathField = createMathField(mount, strings);
+            const mathField = createMathField(mount, mockStrings);
             // create fraction and move cursor to denominator
             mathField.cmd("\\frac");
             mathField.keystroke("Down");
@@ -78,7 +78,7 @@ describe("MathQuill Helpers", () => {
         it("returns IN_SUB_SCRIPT when in subscript", () => {
             // Arrange
             const mount = document.createElement("div");
-            const mathField = createMathField(mount, strings);
+            const mathField = createMathField(mount, mockStrings);
             // "_" triggers a subscript
             mathField.typedText("6_");
 
@@ -92,7 +92,7 @@ describe("MathQuill Helpers", () => {
         it("returns IN_SUPER_SCRIPT when in superscript", () => {
             // Arrange
             const mount = document.createElement("div");
-            const mathField = createMathField(mount, strings);
+            const mathField = createMathField(mount, mockStrings);
             // "^" triggers a superscript
             mathField.typedText("6^");
