+++ packages/math-input/src/components/keypad/keypad-button.stories.tsx
@@ -2,7 +2,7 @@ import {action} from "@storybook/addon-actions";
 import * as React from "react";
 
 import KeyConfigs from "../../data/key-configs";
-import {strings} from "../../mock-strings";
+import {mockStrings} from "../../strings";
 
 import {KeypadButton} from "./keypad-button";
 
@@ -19,7 +19,7 @@ export default {
     argTypes: {
         keyConfig: {
             control: "select",
-            options: {...KeyConfigs(strings)},
+            options: {...KeyConfigs(mockStrings)},
         },
         tintColor: {
             control: "color",
@@ -61,7 +61,7 @@ export const AllButtons: ComponentStory<typeof KeypadButton> = ({
             gap: "25px",
         }}
     >
-        {Object.keys(KeyConfigs(strings)).map((key) => (
+        {Object.keys(KeyConfigs(mockStrings)).map((key) => (
             <div
                 key={key}
                 style={{
@@ -72,7 +72,7 @@ export const AllButtons: ComponentStory<typeof KeypadButton> = ({
             >
                 {key}
                 <KeypadButton
-                    keyConfig={KeyConfigs(strings)[key]}
+                    keyConfig={KeyConfigs(mockStrings)[key]}
                     onClickKey={action("pressed")}
                     coord={[0, 0]}
                 />
