+++ packages/math-input/src/types.ts
@@ -68,69 +68,3 @@ export type KeypadContextType = {
     ) => void;
     scrollableElement: HTMLElement | null | undefined;
 };
-
-/**
- * The translated strings that are used to render the Math Input.
- */
-export type MathInputStrings = {
-    mathInputBox: string;
-    fingerTap: string;
-    before: ({obj}: {obj: string}) => string;
-    after: ({obj}: {obj: string}) => string;
-    "beginning of": ({obj}: {obj: string}) => string;
-    "end of": ({obj}: {obj: string}) => string;
-    Baseline: string;
-    Superscript: string;
-    selected: ({obj}: {obj: string}) => string;
-    "no answer": string;
-    "nothing selected": string;
-    "nothing to the right": string;
-    "nothing to the left": string;
-    "block is empty": string;
-    "nothing above": string;
-    labelValue: ({label, value}: {label: string; value: string}) => string;
-    plus: string;
-    minus: string;
-    negative: string;
-    times: string;
-    divide: string;
-    decimal: string;
-    percent: string;
-    cdot: string;
-    equalsSign: string;
-    notEqualsSign: string;
-    greaterThanSign: string;
-    lessThanSign: string;
-    greaterThanOrEqualToSign: string;
-    lessThanOrEqualSign: string;
-    fractionExpressionInNumerator: string;
-    fractionExcludingExpression: string;
-    customExponent: string;
-    square: string;
-    cube: string;
-    squareRoot: string;
-    cubeRoot: string;
-    radicalWithCustomRoot: string;
-    leftParenthesis: string;
-    rightParenthesis: string;
-    naturalLog: string;
-    logBase10: string;
-    logCustomBase: string;
-    sine: string;
-    cosine: string;
-    tangent: string;
-    pi: string;
-    theta: string;
-    upArrow: string;
-    rightArrow: string;
-    downArrow: string;
-    leftArrow: string;
-    navOutOfParentheses: string;
-    navOutOfExponent: string;
-    navOutOfBase: string;
-    navIntoNumerator: string;
-    navOutOfNumeratorIntoDenominator: string;
-    navOutOfDenominator: string;
-    delete: string;
-    dismiss: string;
-};
